+++ web_ui/app.py
@@ -13,6 +13,7 @@ from starlette.staticfiles import StaticFiles
 
 from . import deps
 from .routes.dashboard import routes as dashboard_routes
+from .routes.memories import routes as memories_routes
 
 load_dotenv()
 
@@ -38,6 +39,7 @@ app = Starlette(
     routes=[
         Mount("/static", StaticFiles(directory=str(BASE_DIR / "static")), name="static"),
         *dashboard_routes,
+        *memories_routes,
     ],
     lifespan=lifespan,
 )